# fsa-eventfindr-backend/application/api-spec/src/main/resources/openapi/eventfindr.yaml
openapi: 3.0.0
info:
  title: EventfindR API
  version: 1.0.0
  description: >
    REST API pre platformu EventfindR – vyhľadávanie, sledovanie a správa eventov.
    Používatelia môžu prehliadať eventy, organizátori ich pridávať a spravovať.
    Systém podporuje označenie účasti (zúčastním sa / sledujem).
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []

paths:
  /events:
    get:
      summary: Získanie zoznamu všetkých eventov
      operationId: getAllEvents
      tags:
        - Event
      responses:
        '200':
          description: Zoznam eventov
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Vytvorenie nového eventu
      operationId: createEvent
      tags:
        - Event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event úspešne vytvorený
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /events/{id}:
    get:
      summary: Získanie detailu eventu podľa ID
      operationId: getEventById
      tags:
        - Event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Detail eventu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /events/{id}/attend:
    post:
      summary: Označenie účasti na evente (zúčastním sa / sledujem)
      operationId: attendEvent
      tags:
        - Event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendEventRequest'
      responses:
        '201':
          description: Účasť úspešne zaregistrovaná
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict – účasť už existuje
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /users:
    post:
      summary: Vytvorenie nového používateľa
      operationId: createUser
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Používateľ úspešne vytvorený
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string

    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
          description: Názov eventu
        description:
          type: string
          description: Popis eventu
        location:
          type: string
          description: Miesto konania
        eventDate:
          type: string
          format: date-time
          description: Dátum a čas konania
        price:
          type: number
          format: double
          description: Cena vstupného
        ticketUrl:
          type: string
          description: Odkaz na kúpu lístkov
        imageUrl:
          type: string
          description: URL obrázku eventu
        genre:
          type: string
          description: Žáner eventu
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
          description: Stav eventu
        featured:
          type: boolean
          description: Či je event zvýraznený
        canceled:
          type: boolean
          description: Či je event zrušený
        attendingCount:
          type: integer
          readOnly: true
        watchingCount:
          type: integer
          readOnly: true
        commentCount:
          type: integer
          readOnly: true
        capacity:
          type: integer
          description: Maximálna kapacita (null = neobmedzená)
        averageRating:
          type: number
          format: double
          readOnly: true
        ratingCount:
          type: integer
          readOnly: true
        artists:
          type: array
          items:
            $ref: '#/components/schemas/EventArtist'
        created:
          type: string
          format: date-time
          readOnly: true
        organizer:
          $ref: '#/components/schemas/User'
          readOnly: true

    CreateEventRequest:
      type: object
      required:
        - name
        - location
        - eventDate
      properties:
        name:
          type: string
          minLength: 1
          pattern: "^.*\\S.*$"
          description: Názov eventu
        description:
          type: string
          description: Popis eventu
        location:
          type: string
          minLength: 1
          pattern: "^.*\\S.*$"
          description: Miesto konania
        eventDate:
          type: string
          format: date-time
          description: Dátum a čas konania
        price:
          type: number
          format: double
          description: Cena vstupného (voliteľné)
        ticketUrl:
          type: string
          description: Odkaz na kúpu lístkov (voliteľné)
        imageUrl:
          type: string
          description: URL obrázku eventu (voliteľné)
        genre:
          type: string
          description: Žáner eventu
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
          description: Stav eventu (DRAFT alebo PUBLISHED)
        capacity:
          type: integer
          description: Maximálna kapacita (voliteľné)
        artists:
          type: array
          items:
            $ref: '#/components/schemas/EventArtistRequest'

    EventsResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'

    AttendEventRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/AttendanceStatus'

    AttendanceStatus:
      type: string
      enum:
        - ATTENDING
        - WATCHING

    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        email:
          type: string
          format: email
        organizationName:
          type: string
          description: Názov organizácie (len pre organizátorov)
        organizationDescription:
          type: string
          description: Popis organizácie (len pre organizátorov)
        artistName:
          type: string
          description: Umelecké meno (len pre umelcov)
        artistDescription:
          type: string
          description: Popis umelca (len pre umelcov)

    EventArtist:
      type: object
      properties:
        id:
          type: integer
          format: int64
        artistUserId:
          type: integer
          format: int64
        artistName:
          type: string
        sortOrder:
          type: integer

    EventArtistRequest:
      type: object
      required:
        - artistName
      properties:
        artistUserId:
          type: integer
          format: int64
        artistName:
          type: string
          minLength: 1

    CreateUserRequest:
      type: object
      required:
        - name
        - email
        - role
      properties:
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        email:
          type: string
          format: email

    UserRole:
      type: string
      enum:
        - USER
        - ORGANIZER
        - ARTIST
        - ADMIN

    EventComment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        eventId:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User'
        content:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        created:
          type: string
          format: date-time

    Notification:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        eventId:
          type: integer
          format: int64
        type:
          type: string
          enum:
            - EVENT_REMINDER
            - EVENT_CANCELED
            - NEW_COMMENT
            - NEW_FOLLOWER
            - NEW_EVENT
            - EVENT_UPDATED
        message:
          type: string
        read:
          type: boolean
        created:
          type: string
          format: date-time